Added some genotype tests for BCF2, which all pass.  Of course that's because I commented out the ones that didn't

+++ public/java/src/org/broadinstitute/sting/utils/codecs/bcf2/BCF2Decoder.java
@@ -160,7 +160,7 @@ public class BCF2Decoder {
                 case INT8:
                 case INT16:
                 case INT32: return value;
-                case FLOAT: return (double)rawFloatToFloat(value);
+                case FLOAT: return rawFloatToFloat(value);
                 case CHAR:  return value & 0xFF; // TODO -- I cannot imagine why we'd get here, as string needs to be special cased
                 default:    throw new ReviewedStingException("BCF2 codec doesn't know how to decode type " + type );
             }
